requirements.txt
psutil
opencv-python
torch
torchvision
pyserial
numpy